pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Arina Zubova <arina777pototskaya@gmail.com>"]
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
Jinja2 = "^3.1.2"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
